Make sure matching is case-sensitive in org-caldav-get-uid.
If case-fold-search is t, the matching done to chop off possible
prefixes added by org-mode (i.e. the [A-Z][A-Z][0-9]*-) regexp, may
match a legitimate UID that just happens to have the first chunk
be two characters and followed by only numbers. This results in the
first chunk of the UID being chopped off, breaking UID matching.

This makes sure case-fold-search is set to nil in org-caldav-get-uid.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>