Add support for OAuth2
This adds support for OAuth2 authentication, currently only for Google
Calendar. Adapt test suite as necessary. This also means that we need
to do all the actual retrieving from URLs ourselves and not leave it
to url-dav, so we can switch to OAuth2 as necessary.

This feature requires the user to register its own application at the
Google Developer Console, since I'm not allowed to distribute my
own. Also, the OAuth2 library must be installed from GNU ELPA.