Detect if org files were removed between syncs (issue #113)
Save org-caldav-files alongside the sync state, so that we can detect
if files were removed from that list. If so, ask the user whether she
really likes to sync anyway. If a file was really removed, this would
delete all entries from the calendar that were in that file, since
org-caldav will see them as deleted. Mention this in Readme.

Add test for this.