Skip VTODO entries on CalDAV server
On CalDAV servers that mix Calendar and TODO entries, org-caldav syncs an empty
"No Title" entry with no UID for every TODO encountered on the server.  These
entries are not cached in the org-caldav sync db, so they are duplicated every
time you resync.  This can quickly lead to thousands of dummy entries in your
org inbox.

This patch just skips VTODO entries entirely when they are encountered.  The
UID is logged in the eventdb with an `ignored` status, and those UIDs are
skipped when checking for changes on re-sync.