Fix bug where link gets deleted from description
In ox-icalendar export, a link in the description looks like
"<https://orgmode.org>".

However, this matches the regex we were previously using to filter out
timestamps from the description. (Generally, ox-icalendar doesn't
remove a leading timestamp from the description, unless it's SCHEDULED
or DEADLINE).

This wasn't a problem before because most entries would have a
timestamp that would be filtered out, but with the new VTODO support
we now can have unscheduled todo entries, and if such an entry's
description contains links, the first one will be removed.

Fix the problem by making the timestamp regex more specific. Also, add
a unit test to check that initial timestamps are still filtered out,
but links are not.

Aside: I was worried this would be complicated by diary-sexp
timestamps, but luckily these aren't currently exported by
ox-icalendar, so we don't need to worry about that for now. E.g., see:
https://list.orgmode.org/8735bcv3xj.fsf@localhost/